app/requirements.txt
Django==3.2.16
gunicorn==20.1.0
psycopg2-binary==2.9.5
redis==4.3.4
django-redis==5.2.0
channels==3.0.4
channels-redis==3.3.1
django-cors-headers==3.10.0
django-leaflet==0.28.2
djangorestframework==3.11.1
drf-extra-fields==3.2.1
numpy==1.21.3
opencv-python==4.6.0.66
Pillow==8.4.0